Add conditional logic for non sandbox deployment
When NOT in the sandbox we do not create the hostedzone, that is part of the project onboarding process. Due to the need for a declaired dependancy between the hostedzone and the records there is aneed to declair the records twice, once when creating the hostedzone and once when deploying into an automated environment.